requirements.txt
streamlit==1.25.0
pandas==2.2.2
spacy==3.6.0
scikit-learn==1.3.2
python-docx==0.8.11
pdfplumber==0.10.0
streamlit
spacy
pdfplumber
python-docx
scikit-learn
